requirements.txt
annotated-types==0.7.0
asarPy==1.0.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Cython==3.0.12
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
Mako==1.3.9.dev0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.27.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.52.2
pyparsing==3.2.4
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
Flask==2.3.5
google-generativeai==0.8.5
pyttsx3==2.90
